requirements.txt
Django
djangorestframework
djoser
nose
black